requirements.txt
streamlit
flask
flask-cors
openai-whisper
transformers
spacy
scikit-learn
reportlab
pypdf2
yake
gtts
pytesseract
pdf2image
pillow
cryptography
python-dotenv
sumy
prometheus-client
